tests/wpt/tests/navigation-api/navigation-activation/activation-replace.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  let before_key = i.contentWindow.navigation.currentEntry.key;
  let before_id = i.contentWindow.navigation.currentEntry.id;
  let before_url = i.contentWindow.navigation.currentEntry.url;
  i.contentWindow.navigation.navigate("/common/blank.html?a", { history: "replace" });
  await new Promise(resolve => i.onload = () => t.step_timeout(resolve, 0));
  let current_entry_after_nav = i.contentWindow.navigation.currentEntry;
  let from_entry_after_nav = i.contentWindow.navigation.activation.from;

  // activation.entry is the current entry. activation.from is a new
  // NavigationHistoryEntry object (because there is a new window), with the
  // same key/id/url as before the navigation, and an index of -1 because it is
  // not present in the entries array.
  assert_equals(i.contentWindow.navigation.entries().length, 1);
  assert_equals(i.contentWindow.navigation.activation.entry, current_entry_after_nav);
  assert_equals(i.contentWindow.navigation.activation.entry.index, 0);
  assert_equals(from_entry_after_nav.key, before_key);
  assert_equals(from_entry_after_nav.id, before_id);
  assert_equals(from_entry_after_nav.url, before_url);
  assert_equals(from_entry_after_nav.index, -1);
  assert_equals(i.contentWindow.navigation.activation.navigationType, "replace");

  await i.contentWindow.navigation.navigate("/common/blank.html?a#fragment", { history: "replace" }).finished;

  // activation.entry is no longer navigation.currentEntry and is disposed.
  // activation.from has not changed.
  assert_equals(i.contentWindow.navigation.entries().length, 1);
  assert_equals(i.contentWindow.navigation.activation.entry, current_entry_after_nav);
  assert_not_equals(i.contentWindow.navigation.activation.entry,
                    i.contentWindow.navigation.currentEntry);
  assert_equals(i.contentWindow.navigation.activation.entry.index, -1);
  assert_equals(i.contentWindow.navigation.activation.from, from_entry_after_nav);
  assert_equals(i.contentWindow.navigation.activation.navigationType, "replace");
}, "navigation.activation after replace");
</script>
