LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-activation/activation-replace-expected.txt



FAIL navigation.activation after replace assert_equals: expected 1 but got 2
